add loading

+++ src/views/front/CartView.vue
@@ -1,7 +1,9 @@
 <template>
     <div class="bg-[#070707f0] text-white">
         <NavBar/>
-
+        <div v-if="isLoading" class="fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center">
+            <img class="h-16 w-16" src="/Rolling-1s-150px.gif" alt="loading">
+        </div>
         <header>
             <div class="h-[500px] bg-cover bg-center flex flex-col items-center justify-center text-center opacity-85" style="background-image: url(https://images.unsplash.com/photo-1552844418-3d618ca9af68?q=80&w=2076&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D);">
                 <div class="">
@@ -12,7 +14,7 @@
             </div>
         </header>
         <main class="container mx-auto p-6 md:p-24 mb-6">
-            <table class="border-t-2 w-full py-8">
+            <table v-if="info !== null && info.carts.length !== 0" class="border-t-2 w-full py-8">
                 <thead class="bg-[#272626e8] border-b-2 text-center">
                     <tr>
                         <th class="p-4"></th>
@@ -31,14 +33,21 @@
                         <td class="w-[14%]">{{ item.product.title }}</td>
                         <td class="w-[14%]">{{ item.product.price }}</td>
                         <td class="">
-                            <div class="w-[80%] border-2 rounded-lg mb-4 md:mb-0">
-                                <button :disabled="item.qty === 1" @click="minusNum(item)" class="duration-500 hover:bg-white hover:text-black font-extrabold text-white py-2 rounded-l-lg w-[20%] text-center" type="button">-</button>
-                                <input :value="item.qty" type="text" class=" text-center bg-gray-800  p-2 w-[60%]">
-                                <button @click="addNum(item)" class="duration-500 hover:bg-white hover:text-black font-extrabold text-white py-2 rounded-r-lg w-[20%] text-center" type="button">+</button>
+                            <div class="w-[80%] border-2 rounded-lg mb-4 md:mb-0 flex items-center">
+                                <button :disabled="item.qty === 1" @click="minusNum(item, index)" class="duration-500 hover:bg-white hover:text-black font-extrabold text-white py-2 rounded-l-lg w-[20%] text-center" type="button">-</button>
+                                <img v-if="isChangeNum && itemIndex === index" class="h-6 w-6 mx-auto " src="/Rolling-1s-150px.gif" alt="loading">
+                                <input @change="changeNum($event, item, index)" v-else :value="item.qty" type="text" class=" text-center bg-gray-800  p-2 w-[60%]">
+                                
+                                <button @click="addNum($event, item, index)" class="duration-500 hover:bg-white hover:text-black font-extrabold text-white py-2 rounded-r-lg w-[20%] text-center" type="button">+</button>
                             </div>
                         </td>
                         <td class="w-[10%]">{{ item.total }}</td>
-                        <td @click="deleteItem(item.id)" class="w-[10%]">垃圾桶</td>
+                        <td  class="w-[10%]">
+                            <button @click="deleteItem(item.id)" type="button">
+                                <svg class="w-6 h-6" fill="white" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M170.5 51.6L151.5 80h145l-19-28.4c-1.5-2.2-4-3.6-6.7-3.6H177.1c-2.7 0-5.2 1.3-6.7 3.6zm147-26.6L354.2 80H368h48 8c13.3 0 24 10.7 24 24s-10.7 24-24 24h-8V432c0 44.2-35.8 80-80 80H112c-44.2 0-80-35.8-80-80V128H24c-13.3 0-24-10.7-24-24S10.7 80 24 80h8H80 93.8l36.7-55.1C140.9 9.4 158.4 0 177.1 0h93.7c18.7 0 36.2 9.4 46.6 24.9zM80 128V432c0 17.7 14.3 32 32 32H336c17.7 0 32-14.3 32-32V128H80zm80 64V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16zm80 0V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16zm80 0V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16z"/></svg>
+                            </button>
+                            
+                        </td>
                     </tr>
                 </tbody>
                 <tfoot class="text-right">
@@ -48,7 +57,7 @@
                 </tfoot>
             </table>
                 
-            <div class="py-8 text-center flex flex-col md:flex-row items-center justify-between">
+            <div v-if="info !== null && info.carts.length !== 0" class="py-8 text-center flex flex-col md:flex-row items-center justify-between">
                 <router-link to="/products" class="border-2 p-2 rounded-lg duration-500 hover:bg-white hover:text-black w-[50%] md:w-[20%] mb-4">
                     繼續購物
                 </router-link>
@@ -57,9 +66,9 @@
                 </router-link>
             </div>
 
-            <div class="py-8">
-                <p class="p-2">目前沒有商品</p>
-                <router-link to="/products" class="border-2 p-2 rounded-lg duration-500 hover:bg-white hover:text-black w-[50%] md:w-[20%] mb-4">
+            <div v-if="info !== null && info.carts.length === 0" class="py-8 text-center">
+                <p class="p-2 text-3xl mb-8">尚未加入商品</p>
+                <router-link to="/products" class="inline-block border-2 p-2 rounded-lg duration-500 hover:bg-white hover:text-black w-[50%] md:w-[20%] mb-4">
                     前往購物
                 </router-link>
             </div>
@@ -85,9 +94,14 @@
 
     const router = useRouter()
 
-    const info = ref('')
+    const isLoading = ref(true)
+    const isChangeNum = ref(false)
+    const itemIndex = ref('')
+    const info = ref(null)
 
-    const addNum = async (item) => {
+    const addNum = async (item, index) => {
+        itemIndex.value = index
+        console.log(itemIndex.value, index)
         item['qty'] +=1
         const data = {
             "data": {
@@ -95,18 +109,20 @@
                 "qty": item['qty']
             }
         }
-
+        isChangeNum.value = true
         try {
             const res = await updateCartItem(item['id'], data)
             console.log(res)
-            const res2 = await getCartInfo()
-            info.value = res2.data
+            const cartInfo = await getCartInfo()
+            isChangeNum.value = false
+            info.value = cartInfo.data
         } catch (error) {
             console.log(item)
         }
     }
 
-    const minusNum = async (item) => {
+    const minusNum = async (item, index) => {
+        itemIndex.value = index
         item['qty'] === 1 ? null : item['qty'] -=1
         const data = {
             "data": {
@@ -114,26 +130,77 @@
                 "qty": item['qty']
             }
         }
-
+        isChangeNum.value = true
         try {
             const res = await updateCartItem(item['id'], data)
             console.log(res)
-            const res2 = await getCartInfo()
-            info.value = res2.data
+            const cartInfo = await getCartInfo()
+            isChangeNum.value = false
+            info.value = cartInfo.data
         } catch (error) {
             console.log(error)
         }
     }
 
-    const deleteItem = async (cartId) => {
+    const changeNum = async (e, item, index) => {
+        itemIndex.value = index
+        let num = parseInt(e.target.value)
+        const data = {
+            "data": {
+                "product_id": item['product_id'],
+                "qty": num
+            }
+        }
+        console.log(data)
+        isChangeNum.value = true
         try {
-            const res = await deleteCartItem(cartId)
+            const res = await updateCartItem(item['id'], data)
             console.log(res)
-            const res2 = await getCartInfo()
-            info.value = res2.data
+            const cartInfo = await getCartInfo()
+            isChangeNum.value = false
+            info.value = cartInfo.data
         } catch (error) {
             console.log(error)
         }
+
+    }
+
+    const deleteItem =  (cartId) => {
+        Swal.fire({
+          title: "確定刪除？",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "確定",
+          cancelButtonText: "取消",
+          
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            Swal.fire({
+                title: "正在刪除",
+                didOpen: () => {
+                    Swal.showLoading()
+                },
+            })
+            try {
+                const res = await deleteCartItem(cartId)
+                console.log(res)
+                const cartInfo = await getCartInfo()
+                info.value = cartInfo.data
+                Swal.fire({
+                title: "刪除成功",
+                icon: "success",
+                didOpen: () => {
+                    Swal.hideLoading()
+                },
+            })
+            } catch (error) {
+                console.log(error)
+            }
+          }
+        })
+        
         
     }
 
@@ -141,6 +208,7 @@
     onMounted( async () => {
         try {
             const res = await getCartInfo()
+            isLoading.value = false
             console.log(res)
             info.value = res.data
         } catch (error) {
